requirements.txt
certifi==2022.12.7
chardet==3.0.4
click==6.7
colorama==0.3.9
dataclasses==0.8
Flask==1.1.1
idna==2.6
itsdangerous==0.24
Jinja2==3.0.3
lxml==4.1.1
MarkupSafe==2.0.1
numpy==1.12.1
pandas==0.23.3
Pillow==5.2.0
python-dateutil==2.6.1
python-docx==0.8.11
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==2.0.3